+++ src/main/kotlin/com/relogging/server/oauth/service/PrincipalOAuthUserService.kt
@@ -1,18 +1,18 @@
-package com.relogging.server.oauth
+package com.relogging.server.oauth.service
 
 import com.relogging.server.domain.user.entity.SocialType
 import com.relogging.server.global.exception.GlobalErrorCode
 import com.relogging.server.global.exception.GlobalException
 import com.relogging.server.oauth.provider.GoogleUserInfo
 import com.relogging.server.oauth.provider.KakaoUserInfo
-import com.relogging.server.oauth.provider.OAuth2UserInfo
+import com.relogging.server.oauth.provider.OAuthUserInfo
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest
 import org.springframework.security.oauth2.core.user.OAuth2User
 import org.springframework.stereotype.Service
 
 @Service
-class OAuth2UserService : DefaultOAuth2UserService() {
+class PrincipalOAuthUserService : DefaultOAuth2UserService() {
     override fun loadUser(userRequest: OAuth2UserRequest): OAuth2User {
         val oAuth2User: OAuth2User = super.loadUser(userRequest)
         println("Attributes: ${oAuth2User.attributes}")
@@ -20,9 +20,9 @@ class OAuth2UserService : DefaultOAuth2UserService() {
         val socialType: SocialType = this.getSocialType(userRequest)
         println("SocialType: $socialType")
 
-        val oAuth2UserInfo: OAuth2UserInfo =
+        val oAuthUserInfo: OAuthUserInfo =
             this.getOAuth2UserInfo(socialType, oAuth2User.attributes)
-        println("OAuth2UserInfo: $oAuth2UserInfo")
+        println("OAuth2UserInfo: $oAuthUserInfo")
 
         return oAuth2User
     }
@@ -40,7 +40,7 @@ class OAuth2UserService : DefaultOAuth2UserService() {
     fun getOAuth2UserInfo(
         socialType: SocialType,
         attributes: Map<String, Any>,
-    ): OAuth2UserInfo {
+    ): OAuthUserInfo {
         return when (socialType) {
             SocialType.GOOGLE -> GoogleUserInfo(attributes)
             SocialType.KAKAO -> KakaoUserInfo(attributes)
